refactor: Introduced systems for tiles, gizmo subtypes, and tickable gizmos.
Model contains 20x20 tiles. A tile can contain a gizmo which is an abstract type. Tile implements
Drawable which is used by the view to draw the gizmo inside the tile if it exists. There is
currently no collision detection on the flipper, and the ball is not implemented into the ticking
system.