feat: Flipper reacts with key up/down correctly
When correct key is help down, flipper stays flipped until key is released. Gizmo now implements
GizmoEventListener which supplies interface for keyUp, keyDown, and genericTrigger for when the
event is not accociated with a key. When the is a key event, model now finds the accociated
GizmoEventListener using a dictionary, then calls the correct function.